Setup the enviroment
Add an index.html file with the basic template for a page, a
style.css file in a styles folder and a script.js file in a
scripts folder.

These are all the necessary files to develop the project and the
basic template is the starting point of the html code for the
page.