Create a 16x16 grid of square divs
Add a container div in html, a grid generating script with square
divs in the container and styles for the html, body, main elements
and .container,.cell classes.

The container div will be used to append the divs generated by
the script to create the canvas. It is implemented in html because
it will be fixed, with only the grid changing. The styles for the
html, body and main are to make the webpage responsive, while the
styles for both classes are create the layout of the grid.