Implement a reset button for the grid
Add a button with a onclick call to resetGrid in index.html and
increase the canvas size to 960px. Refactor the previous code in
a createGrid function and add a getGridSize, a clearGrid and
a resetGrid function.

These are all used to implement the reset button for the grid as
required by the project. Moreover, The refactor helps reutilize
the previous code in creating new grids, the getGridSize allows
to get the right input for the grid size, the clearGrid removes
the previous grid and the resetGrid start the grid resetting
process.