Implement random color trail effect
Add black and random buttons in index.html to change coloring
modes, a coloringMode variable to keep track of the modes and the
onclick call functions for each button. Rework the changeColor
function to work acording to the selected mode.

The random color trail effect is an optional requirement of the
project. Moreover the changes aim to improve the code organization
and adapt for future color mode implementations.