Implement color palette for fixed color mode
Add a color input and rename the buttons in index.html, add a
variable to store the current color, create const variables to
store the main html elements, add a onchange call function to
change the current color and rework the changeColor function to
use the color palette.

This is an additional feature of the project that amplifies its
usability in drawing.