Style the page
Add a fontawesome script to use the github symbol, add a header
with a h1 title, add two divs to group buttons above and below
the canvas and add a footer to show my informations. Rework the
buttons on change call functions to highlight the active mode and
rename the incrementButton to shadeButton. Style the page with
a predefined color palette, keeping the header and footer "fixed",
aligning the elements horizontally and vertically, adding buttons
effects and media queries for landscape mobile and bigger screens.

Although not required by the project, a responsive style improves
user interaction for most plataforms.